+++ components/LocationSingle/CampusInfo/index.js
@@ -1,17 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { capitalize, find } from 'lodash';
+import { find } from 'lodash';
 
-import { Box, Button, Cell, Divider, Icon, utils } from 'ui-kit';
+import { Box, Button, Cell, utils } from 'ui-kit';
 
 import { campusLinks } from '../../../lib/locationData';
 import Styled from '../LocationSingle.styles';
-import CfEverywhereButtons from './CFEverywhereButtons';
 import PastorCard from './PastorCard';
 import { validDaysOfWeek } from '../utils';
-import { format } from 'date-fns';
 
-const StyledDivider = props => <Divider bg="secondarySubdued" {...props} />;
+import {
+  TrinityButtons,
+  CfEverywhereButtons,
+  WeekdayScheduleDisplay,
+} from './customComponents';
+import { is } from 'date-fns/locale';
 
 const CampusInfo = ({
   name,
@@ -93,106 +96,34 @@ const CampusInfo = ({
                 ))}
               </Styled.InfoBox>
             )}
-            {/* Address and Church You Call Home */}
+            {/* Weekday Schedule / Custom Campus Info */}
+
+            {/* Custom Info for CF Everywhere and Trinity */}
             {name === 'Online (CF Everywhere)' && <CfEverywhereButtons />}
-            {name === 'Trinity' || name === 'Online (CF Everywhere)' ? (
-              <Box ml="base" display={{ _: 'none', md: 'flex' }} my="l">
-                <Box flex="1">
-                  <Box as="h3" pr="xl" color="secondary" maxWidth={200}>
-                    {name === 'Trinity'
-                      ? 'Experience Something New!'
-                      : 'Groups & Classes'}
-                  </Box>
-                </Box>
-                {name === 'Trinity' ? (
-                  <Box flex="2">
-                    Have you been searching for a meaningful community but
-                    haven’t found it yet? If so, you’re not alone. Trinity
-                    Church by Christ Fellowship is a new church experience
-                    coming to your neighborhood! This community location in Palm
-                    Beach Gardens offers a different way to experience church so
-                    you can get to know people in your neighborhood and enjoy a
-                    more traditional worship setting. Find a place for you and
-                    your family to belong with even more regional events offered
-                    just down the street—it’s big church made small, and you’ll
-                    feel right at home!
-                  </Box>
-                ) : (
-                  <Box ml={!isWeekdaySchedule && 'base'} flex="2">
-                    <Box>
-                      Discover groups and classes to help you grow in your
-                      relationship with God and others.
-                    </Box>
-                    <Button size="s" mt="s">
-                      Find a Group or Class
-                    </Button>
-                  </Box>
-                )}
-              </Box>
-            ) : null}
-            {isWeekdaySchedule && [
-              <Box display={{ _: 'none', md: 'flex' }} my="l">
-                <Box flex="1">
-                  <Box as="h3" pr="xl" color="secondary" maxWidth={200}>
-                    During the Week
-                  </Box>
-                </Box>
-                <Box
-                  flex="2"
-                  display="grid"
-                  gridColumnGap="xs"
-                  gridRowGap="l"
-                  gridTemplateColumns="repeat(2, 1fr)"
-                >
-                  {validDaysOfWeek(weekdaySchedules)?.map(day => {
-                    return (
-                      <Box>
-                        <Box as="h3" mb="xs">
-                          {capitalize(Object?.keys(day))}
-                        </Box>
-                        {day[Object?.keys(day)]?.map(event => {
-                          const formattedTime = format(
-                            //parse time into valid date to format(added random day for date to parse)
-                            Date.parse(`2023-01-01T${event?.time}`),
-                            'h:mm a'
-                          );
-                          return (
-                            <Box display="flex" alignItems="center">
-                              {`${formattedTime} - 
-                                ${event?.title}`}
-                              {event?.url && (
-                                <Box mb="3px" as="a" href={event?.url}>
-                                  <Icon ml="3px" size={13} name="link-out" />
-                                </Box>
-                              )}
-                            </Box>
-                          );
-                        })}
-                      </Box>
-                    );
-                  })}
-                </Box>
-              </Box>,
-              <StyledDivider
-                display={{ _: 'none', md: 'flex' }}
-                width="100%"
-              />,
-            ]}
-            {name !== 'Online (CF Everywhere)' && (
-              <Box display={{ _: 'none', md: 'flex' }} mt="l">
-                {isWeekdaySchedule && <Box flex="1" />}
-                <Box ml={!isWeekdaySchedule && 'base'} flex="2">
-                  <Box as="h3">Additional Groups & Classes</Box>
-                  <Box>
-                    Discover groups and classes to help you grow in your
-                    relationship with God and others.
-                  </Box>
-                  <Button size="s" mt="s">
-                    Find a Group or Class
-                  </Button>
-                </Box>
-              </Box>
+            {name === 'Trinity' && <TrinityButtons />}
+
+            {/* Weekday Schedule */}
+            {isWeekdaySchedule && (
+              <WeekdayScheduleDisplay weekdaySchedules={weekdaySchedules} />
             )}
+
+            {/* Groups and Classes Section */}
+            <Box display="flex" mt="l" px="base">
+              <Box
+                textAlign={{ _: 'center', md: 'left' }}
+                ml={!isWeekdaySchedule && 'base'}
+                flex="2"
+              >
+                <Box as="h3">Groups & Classes</Box>
+                <Box>
+                  Discover groups and classes to help you grow in your
+                  relationship with God and others.
+                </Box>
+                <Button size="s" mt="s">
+                  Find a Group or Class
+                </Button>
+              </Box>
+            </Box>
           </Box>
         </Box>
         {/* Campus Pastors */}
@@ -205,6 +136,15 @@ const CampusInfo = ({
           mapLink={mapLink}
         />
       </Cell>
+
+      {isWeekdaySchedule && (
+        <Box mt="xxl" display={{ _: 'inline', md: 'none' }}>
+          <WeekdayScheduleDisplay
+            isMobile
+            weekdaySchedules={weekdaySchedules}
+          />
+        </Box>
+      )}
     </Box>
   );
 };
